+++ aoc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Revision history for aoc
 
+## 0.2.0.0 -- 2025-03-26
+
+* Removed shortestDistanceOnMagma and shortestPathsOnMagma
+* Added shortestDistanceWith and shortestPathsWith
+* Removed Enum constraints on neighbours8, neighbours26, and all enumeration functions.
+* Added generalized n-dimensional taxicab function
+* Added enumerate functions that return pairs instead of triples
+* Fixed floodFill and floodFillWith
+* Fixed choose and permute
+* Added pretty print functions
+* Added basic list operations (takeEvery and chunk)
+* Added traceSleep
+* Added memoization for functions with multiple arguments
+
 ## 0.1.0.1 -- 2025-03-26
 
 * Exported HashMap and Set types to prevent conflicts in type signatures